+++ dejafu/Test/DejaFu/Settings.hs
@@ -25,6 +25,29 @@ module Test.DejaFu.Settings
   , swarmy
 
   -- *** Schedule bounds
+
+  -- | Schedule bounding is used by the 'systematically' approach to
+  -- limit the search-space, which in general will be huge.
+  --
+  -- There are three types of bound:
+  --
+  --  * The 'PreemptionBound', which bounds the number of pre-emptive
+  --    context switches.  Empirical evidence suggests @2@ is a good
+  --    value for this, if you have a small test case.
+  --
+  --  * The 'FairBound', which bounds the difference between how many
+  --    times threads can yield.  This is necessary to test certain
+  --    kinds of potentially non-terminating behaviour, such as
+  --    spinlocks.
+  --
+  --  * The 'LengthBound', which bounds how long a test case can run,
+  --    in terms of scheduling decisions.  This is necessary to test
+  --    certain kinds of potentially non-terminating behaviour, such
+  --    as livelocks.
+  --
+  -- Schedule bounding is not used by the non-systematic exploration
+  -- behaviours.
+
   , Bounds(..)
   , PreemptionBound(..)
   , FairBound(..)
@@ -36,11 +59,75 @@ module Test.DejaFu.Settings
   , noBounds
 
   -- ** The @MemType@
+
+  -- | When executed on a multi-core processor some @CRef@ / @IORef@
+  -- programs can exhibit \"relaxed memory\" behaviours, where the
+  -- apparent behaviour of the program is not a simple interleaving of
+  -- the actions of each thread.
+  --
+  -- __Example:__ This is a simple program which creates two @CRef@s
+  -- containing @False@, and forks two threads.  Each thread writes
+  -- @True@ to one of the @CRef@s and reads the other.  The value that
+  -- each thread reads is communicated back through an @MVar@:
+  --
+  -- > >>> :{
+  -- > let relaxed = do
+  -- >       r1 <- newCRef False
+  -- >       r2 <- newCRef False
+  -- >       x <- spawn $ writeCRef r1 True >> readCRef r2
+  -- >       y <- spawn $ writeCRef r2 True >> readCRef r1
+  -- >       (,) <$> readMVar x <*> readMVar y
+  -- > :}
+  --
+  -- We see something surprising if we ask for the results:
+  --
+  -- > >>> autocheck relaxed
+  -- > [pass] Never Deadlocks
+  -- > [pass] No Exceptions
+  -- > [fail] Consistent Result
+  -- >     (False,True) S0---------S1----S0--S2----S0--
+  -- >
+  -- >     (False,False) S0---------S1--P2----S1--S0---
+  -- >
+  -- >     (True,False) S0---------S2----S1----S0---
+  -- >
+  -- >     (True,True) S0---------S1-C-S2----S1---S0---
+  -- > False
+  --
+  -- It's possible for both threads to read the value @False@, even
+  -- though each writes @True@ to the other @CRef@ before reading.
+  -- This is because processors are free to re-order reads and writes
+  -- to independent memory addresses in the name of performance.
+  --
+  -- Execution traces for relaxed memory computations can include
+  -- \"C\" actions, as above, which show where @CRef@ writes were
+  -- explicitly /committed/, and made visible to other threads.
+  --
+  -- However, modelling this behaviour can require more executions.
+  -- If you do not care about the relaxed-memory behaviour of your
+  -- program, use the 'SequentialConsistency' model.
+
   , MemType(..)
   , defaultMemType
   , lmemtype
 
   -- ** Discard functions
+
+  -- | Sometimes we know that a result is uninteresting and cannot
+  -- affect the result of a test, in which case there is no point in
+  -- keeping it around.  Execution traces can be large, so any
+  -- opportunity to get rid of them early is possibly a great saving
+  -- of memory.
+  --
+  -- A discard function, which has type @Either Failure a -> Maybe
+  -- Discard@, can selectively discard results or execution traces
+  -- before the schedule exploration finishes, allowing them to be
+  -- garbage collected sooner.
+  --
+  -- __Note:__ The predicates and helper functions in Test.DejaFu come
+  -- with discard functions built in, to discard results and traces
+  -- wherever possible.
+
   , Discard(..)
   , ldiscard